[Bug] Fix some definition errors
# Conflicts:
#	cocos/spine/lib/spine-core.d.ts
#	package.json
