+++ scripts/test-declarations/test.d.ts
@@ -1,0 +1,2 @@
+/// <reference path="../../bin/.declarations/cc.d.ts" />
+/// <reference path="../../bin/.declarations/cc.editor.d.ts" />
