+++ scripts/test-declarations/test.js
@@ -1,0 +1,3 @@
+"use strict";
+Object.defineProperty(exports, "__esModule", { value: true });
+Editor.Package;
