+++ scripts/test-declarations/test.ts
@@ -1,0 +1,2 @@
+import { Node } from 'cc';
+Editor.Package;
